Simplifying replace explanation
Again, my opinions and punch-ups:

I think this is too early to give an aside on even the concept of multiple-return values, a concept not used in Javascript, which is presumably the reader's first language.

Also, I think saying "without modifying the thing we passed in" feels more comforting to the reader than saying "instead of modifying the thing we passed in", which reinforces a feeling of "this isn't what you'd expect" instead of "this is how it is". (Although I know this can be a sticking point for many learners!  I think a solid presentation of the actual behavior could bypass this sticky spot entirely!)

Also replaced a couple instances of `thing` with `value` to prevent the reader from needing to context-switch definitions.

Anyways, your tweet got me re-reading `jsforcats` this morning, by all means, take it or leave it, this is very subjective stuff.